Modify apis && models　(Greate renovation)
- `content/github/apis` and `inject/github/apis` files unified in `apis`
- `content/github/models` and `inject/github/models` files unified in `models`
