--- src/content/github/models/Followers.js
@@ -1,55 +1,0 @@
-'use strict'
-
-import Base from './Base'
-
-export default class Followers extends Base {
-  constructor(params){
-    super(params)
-    this.userName = this.params.userName
-    this.userGithubId = 0
-    // override
-    this.type = 'followers'
-  }
-
-  relationships(){
-    return []
-  }
-
-  createId(params){
-    return `${params.userName}-${this.userGithubId}-${this.type}`
-  }
-
-  dataFromOctokit(){
-    return this.authClient.users.listFollowersForUser(
-    {
-      username: `${this.userName}`,
-      per_page: 100,
-      page: 500
-    }).then(result => {
-      this.data = result
-    }).catch(error => {
-      console.log(error)
-    })
-  }
-
-  updateProperties(params = {}){
-    if (params === {}){
-      this.id = this.createId(this.params)
-    } else {
-      Object.keys(params).forEach(key => {
-        let value = params[key]
-        switch(key){
-          case 'userGithubId':
-            this.userGithubId = value
-            this.id = this.createId(this.params)
-            
-            break
-          default:
-            this.id = this.createId(this.params)
-
-            break
-        }
-      })
-    }
-  }
-}
