--- src/inject/apis/main_content/main_li/Issue.js
@@ -1,19 +1,0 @@
-'use strict'
-
-import Base from './Base'
-import RepoLanguage from './RepoLanguage'
-
-export default class Issue extends Base {
-  get name(){
-    return 'issue'
-  }
-
-  get relationships(){
-    return [this.repoLanguage]
-  }
-
-  // private
-  get repoLanguage(){
-    return new RepoLanguage()
-  }
-}
