--- src/inject/apis/main_content/main_li/RepoLanguage.js
@@ -1,9 +1,0 @@
-'use strict'
-
-import Base from './Base'
-
-export default class RepoLanguage extends Base {
-  get name(){
-    return 'repo_language'
-  }
-}
