--- src/inject/models/main_content/main_li/Base.js
@@ -1,13 +1,0 @@
-'use strict'
-
-export default class Base {
-  constructor (id, data) {
-    this.id = id
-    this.data = data
-  }
-
-  setProperties () {
-    let error = new Error('Override in extends class')
-    throw error
-  }
-}
