+++ src/inject/store/actions.js
@@ -1,13 +1,13 @@
 'use strict'
 
 import * as types from './mutation-types'
-import MainLi from '@/inject/apis/main_content/MainLi'
-import Issue from '@/inject/apis/main_content/main_li/Issue'
-import IssueComment from '@/inject/apis/main_content/main_li/IssueComment'
+import StockedComment from '@/apis/github/StockedComment'
+import Issue from '@/models/github/Issue'
+import IssueComment from '@/models/github/IssueComment'
 
 export default {
   fetchDataFromStorage ({ commit, state }) {
-    let api = new MainLi()
+    let api = new StockedComment()
     let issue = new Issue()
     let issuecomment = new IssueComment()
     api.fetchData([issuecomment, issue], (error, payload) => {
