+++ src/inject/store/modules/sidebar/mutation-types.js
@@ -1,0 +1,1 @@
+export const FETCH_LOGIN_USER_DATA = 'FETCH_LOGIN_USER_DATA'
